# web/doc/OpenAPI.yaml
openapi: 3.1.0
info:
  title: ToDo
  description: This is a ToDo application.
  version: 0.0.1
servers:
  - url: http://localhost:{port}
    description: Localhost
    variables:
      port:
        default: "8080"
paths:
...
  /todos/done/{id}:
    get:
      summary: Mark todo item as Done.
      operationId: DoneTodo
      parameters:
        - name: id
          in: path
          description: id of the todo item.
          required: true
          schema:
            type: number
            format: integer
      responses:
        "200":
          description: Shows a table with all todo items
          content:
            text/html:
              schema:
                type: string
  /todos/{id}:
    delete:
      summary: Delete todo item.
      operationId: DeleteTodo
      parameters:
        - name: id
          in: path
          description: id of the todo item.
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shows a table with all todo items
          content:
            text/html:
              schema:
                type: string
...